cleanups for production

+++ public/layouts/main.js
@@ -1,14 +1,25 @@
 $(document).ready(function(){
+    this.response = "";
 	$('#deleteEvent').on('click', function(e){
 		e.preventDefault();
-		var deleteId = $('#deleteEvent').data('delete');
+		myPrompt();
+	});
+    
+    function myPrompt() {
+    if (confirm("Press a button!") === true) {
+        var deleteId = $('#deleteEvent').data('delete');
 		$.ajax({
-			url: '/events/delete/'+deleteId,
+			url: '/jobs/delete/'+deleteId,
 			type:'DELETE',
 			success: function(result){
 				console.log(result);
 			}
 		});
-		window.location = '/events';
-	});
+		window.location = '/myjobs';
+    } else {
+        this.response = 0;
+    }
+}
+    
+    
 });
